Fix issue with drop option
because the line that drops the variables is captured, when there was a problem with a variable, it simply did not drop any variables at all. @bbdaniels , can you please check this is in line with how the command deals with error messages? I believe this is my first edit to iecodebook